Added register, login, delete only admin

+++ Valorant.tv/valobackend/main.py
@@ -1,9 +1,14 @@
 from fastapi import FastAPI, Depends, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
+from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 from sqlalchemy.orm import Session, joinedload
 import models, schemas, database
 from sqlalchemy import or_
 from passlib.context import CryptContext
+from datetime import datetime, timedelta
+from jose import jwt, JWTError
+
+
 
 # Настройка шифровальщика
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
@@ -12,6 +17,41 @@ def get_password_hash(password):
     return pwd_context.hash(password)
 
 
+SECRET_KEY = "SUPER_SECRET_VALORANT_KEY"
+ALGORITHM = "HS256"
+ACCESS_TOKEN_EXPIRE_MINUTES = 60
+
+def create_access_token(data: dict):
+    to_encode = data.copy()
+    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    to_encode.update({"exp": expire})
+    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
+
+def verify_password(plain_password, hashed_password):
+    return pwd_context.verify(plain_password, hashed_password)
+
+oauth_scheme = OAuth2PasswordBearer(tokenUrl="login")
+
+async def get_current_user(token: str = Depends(oauth_scheme), db: Session = Depends(database.get_db)):
+    credentials_exception = HTTPException(
+        status_code= 401,
+        detail="Не удалось проверить учетные данные",
+        headers={"WWW-Authenticate": "Bearer"}
+    )
+    try:
+        # 1. Расшифровываем токен своим SECRET_KEY
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        username: str = payload.get("sub")
+        if username is None:
+            raise credentials_exception
+    except JWTError:
+        # Если токен подделан или просрочен — вылетает ошибка
+        raise credentials_exception
+    # 2. Ищем пользователя в базе
+    user = db.query(models.UserModel).filter(models.UserModel.username==username).first()
+    if user is None:
+        raise credentials_exception
+    return user # Если всё ок, возвращаем объект пользователя
 
 # Создаем таблицы
 models.Base.metadata.create_all(bind=database.engine)
@@ -45,6 +85,22 @@ async def register_user(user_data: schemas.UserCreate, db: Session = Depends(dat
     db.refresh(new_user)
 
     return new_user
+
+@app.post("/login")
+async def login(form_data: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(database.get_db)):
+    user = db.query(models.UserModel).filter(models.UserModel.username==form_data.username).first()
+
+    # 2. Проверяем существование и пароль
+    if not user or not verify_password(form_data.password, user.hashed_password):
+        raise HTTPException(status_code=400, detail="Неверный логин или пароль")
+    
+    # 3. Создаем токен
+    access_token = create_access_token(data={"sub": user.username})
+    return {"access_token": access_token, "token_type": "bearer"}
+
+
+
+
 # --- RANKINGS ---
 @app.get("/rankings/{region}")
 async def get_rankings(region: str, db: Session = Depends(database.get_db)):
@@ -59,12 +115,20 @@ async def create_team(region: str, team: schemas.TeamCreate, db: Session = Depen
     return db_team
 
 @app.delete("/rankings/{team_id}")
-async def delete_team(team_id: int, db: Session = Depends(database.get_db)):
+async def delete_team(
+    team_id: int,
+    db: Session = Depends(database.get_db),
+    current_user: models.UserModel = Depends(get_current_user)    
+):
+    # Проверяем флаг в базе данных
+    if not current_user.is_admin:
+        raise HTTPException(status_code=403, detail="У вас нет прав супер-админа")
+
     team = db.query(models.TeamModel).filter(models.TeamModel.id == team_id).first()
     if not team: raise HTTPException(status_code=404, detail="Team not found")
     db.delete(team)
     db.commit()
-    return {"message": "Deleted"}
+    return {"message": f"Команда удалена администратором {current_user.username}"}
 
 # @app.get("/teams/{team_id}/stats", response_model=)
